pyproject.toml
[project]
name = "investor-agent"
dynamic = ["version"]
description = "A Model Context Protocol server for an investor agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py",
    "hishel>=0.1.3",
    "html5lib>=1.1",
    "httpx>=0.28.1",
    "lxml>=6.0.1",
    "mcp[cli]>=1.14.1",
    "pandas>=2.3.2",
    "pytrends>=4.9.2",
    "requests-cache>=1.2.1",
    "tenacity>=9.1.2",
    "yfinance[nospam]>=0.2.66",
]
authors = [
    { name = "Ferdous", email = "ferdous@hey.com" }
]
keywords = ["investment", "finance", "trading", "agent", "mcp", "model context protocol", "broker", "robinhood", "wallstreetbets"]
classifiers = [
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business :: Financial",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]

[project.optional-dependencies]
ta = [
    "ta-lib>=0.6.7",
]
bridge = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[dependency-groups]
dev = [
    "pydantic-ai>=1.0.10",
    "python-dotenv>=1.1.1",
]

[project.scripts]
investor-agent = "investor_agent.server:mcp.run"
investor-agent-bridge = "investor_agent.bridge:app"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["investor_agent"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "investor_agent/_version.py"
